requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.0.0
opencv-python>=4.5.0

# Optional dependencies for advanced features
tensorboard>=2.7.0
PyQt5>=5.15.0
mss>=6.1.0

# Development dependencies
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0
